move npm out of cni
