+++ npm/parse.go
@@ -3,9 +3,9 @@ package npm
 import (
 	"fmt"
 
-	"github.com/Azure/azure-container-networking/cni/npm/iptm"
-	"github.com/Azure/azure-container-networking/cni/npm/util"
 	"github.com/Azure/azure-container-networking/log"
+	"github.com/Azure/azure-container-networking/npm/iptm"
+	"github.com/Azure/azure-container-networking/npm/util"
 	networkingv1 "k8s.io/api/networking/v1"
 )
