+++ npm/util/util.go
